% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis.R
\name{A006877}
\alias{A006877}
\title{A006877}
\usage{
A006877(n = 1000L, plot = TRUE)
}
\value{
1, 2, 3, 6, 7, 9, 18, 25, 27, 54, 73, 97, 129, 171, 231, 313, 327, 649,
703, 871, 1161, 2223, 2463, 2919, 3711, 6171, 10971, 13255, 17647, 23529,
26623, 34239, 35655, 52527, 77031, 106239, 142587, 156159, 216367, 230631,
410011, 511935, 626331, 837799, ...
}
\description{
A006877: In the "3x+1" problem, these values for the starting value set new
records for number of steps to reach 1.
}
\seealso{
\url{https://oeis.org/A006877}

\code{\link{collatz}}

Other oeis: \code{\link{A133058}}, \code{\link{A265326}}
}
\concept{oeis}
