% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis.R
\name{A133058}
\alias{A133058}
\title{A133058}
\usage{
A133058(n = 2000L, plot = TRUE)
}
\value{
1, 1, 4, 8, 2, 8, 4, 12, 3, 1, 12, 24, 2, 16, 8, 24, 3, 21, 7, 27, 48, 16,
8, 32, 4, 30, 15, 5, 34, 64, 32, 64, 2, 36, 18, 54, 3, 41, 80, 120, 3, 45,
15, 59, 104, 150, 75, 123, 41, 91, 142, 194, 97, 151, 206, 262, 131, 189,
248, 308, 77, 139, 202, 266, 133, 199, 266, 334, 167, ...
}
\description{
A133058: a(0)=a(1)=1; for n>1, a(n) = a(n-1) + n + 1 if a(n-1) and n are
coprime, otherwise a(n) = a(n-1)/gcd(a(n-1),n).
}
\seealso{
\url{https://oeis.org/A133058}

Other oeis: 
\code{\link{A006877}()},
\code{\link{A265326}()}
}
\concept{oeis}
