% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeis.R
\name{A265326}
\alias{A265326}
\title{A265326}
\usage{
A265326(n = 10000L, plot = TRUE)
}
\value{
1, 0, 0, 0, -2, 2, 0, -6, -6, 6, 0, -4, 4, -10, -14, 10, 4, 14, -30, -42,
0, -42, -18, 12, 30, 18, -12, 0, 18, 42, 0, -62, -8, -70, -20, -82, -28,
-34, -62, -8, -26, 8, -62, 62, 34, -28, 8, -28, 28, 62, 82, -8, 98, 28, 0,
-186, -84, -210, -60, ...
}
\description{
A265326: n-th prime minus its binary reversal.
}
\seealso{
\url{https://oeis.org/A265326}

Other oeis: 
\code{\link{A006877}()},
\code{\link{A133058}()}
}
\concept{oeis}
