% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{ch}
\alias{ch}
\title{Calvin and Hobbes}
\usage{
ch(who)
}
\arguments{
\item{who}{choose a favorite}
}
\description{
Draw Calvin and Hobbes
}
\seealso{
\url{http://www.stolaf.edu/people/olaf/cs125/}
}

