% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_plot.r
\name{clustered_heatmap}
\alias{clustered_heatmap}
\title{Generate a heatmap}
\usage{
clustered_heatmap(
  df,
  value_var,
  row_var,
  column_var,
  normalize = "none",
  cluster_rows = F,
  cluster_cols = F,
  row.order = NULL,
  col.order = NULL,
  text = F,
  color_low = "#b35806",
  color_high = "#542788",
  text_size = 8
)
}
\arguments{
\item{df}{a dataframe from which the heatmap should be generated}

\item{value_var}{a string indicating the variable that contains the value for each cell of the heatmap}

\item{row_var}{a string indicating the variable that will be on the rows of the heatmap}

\item{column_var}{a string representing the variable that will be on the columns of the heamtap}

\item{normalize}{one of three strings ("rows", "columns" or "none") indicating which if any dimension to normalize}

\item{cluster_rows}{a boolean indicating if the rows should be clustered}

\item{cluster_cols}{a boolean indicating if the columns should be clustered}

\item{row.order}{r an ordered charactered vector of length of the number of distinct row values in the dataframe representing how that variable should ordered on the plot. The order starts from the bottom of the y-axis.}

\item{col.order}{an ordered charactered vector of length of the number of distinct column values in the dataframe representing how that variable should ordered on the plot. The order starts from the left of the x-axis}

\item{text}{a boolean indicating if text indicating the cell value should be placed in each entry of the heatmap}

\item{color_low}{a string of a color or hexadecimal value indicating the color at the low end of the gradient}

\item{color_high}{a string of a color or a hexadecimal value indicating the color at the high end of the gradient}
}
\value{
a list containing the ggplot heatmap, an ordered character vector of the row order used,and ordered character vector of the column order used
}
\description{
Generate a heatmap
}
