% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{collatz}
\alias{collatz}
\title{Collatz conjecture}
\usage{
collatz(n, stoptime = FALSE)
}
\arguments{
\item{n}{starting number (integer)}

\item{stoptime}{logical; if \code{FALSE} (default), the value at each
iteration is plotted with a summary of the values; if \code{TRUE}, instead,
the first \code{n} stopping times are plotted along with the OEIS sequence
of numbers \code{<= n}, that is, from one to \code{n}, the number of 
iterations each starting value takes to reach 1}
}
\description{
Perform and visualize the Collatz conjecture.
}
\examples{
collatz(19)
collatz(1161)
collatz(1618, stoptime = TRUE)
collatz(5005, stoptime = TRUE)
collatz(63728127)

}
\references{
\url{http://imgs.xkcd.com/comics/collatz_conjecture.png}

\url{http://en.wikipedia.org/wiki/Collatz_conjecture}

\url{https://oeis.org/A006877}
}
