% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_plot.r
\name{dendro_function}
\alias{dendro_function}
\title{perform hierarchical clustering of a dataset}
\usage{
dendro_function(df, row_var, column_var)
}
\arguments{
\item{df}{input dataframe}

\item{id_var}{a string identifying the id variables to be clustered}

\item{measure_var}{a string identifying the measures of the id variable}
}
\value{
an object of class dendrogram
}
\description{
perform hierarchical clustering of a dataset
}
