% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{enigma}
\alias{enigma}
\title{Enigma}
\usage{
enigma(x, i = 1L, prob = NULL, decode = FALSE)
}
\arguments{
\item{x}{a string to encode or decode; alternatively, an object of class
\code{"enigma"}}

\item{i}{indices of the message letters in encoded string}

\item{prob}{optional probabilities of random letters}

\item{decode}{logical; if \code{TRUE}, \code{x} is assumed to be encoded}
}
\description{
Encode and decode a message.
}
\examples{
x <- "this is my secret message. Don't tell anyone"
set.seed(1)
e1 <- enigma(x, c(3, 2, 2, 3))
e1
enigma(e1)
## alternatively
enigma(c(e1), c(3, 2, 2, 3), decode = TRUE)


## use observed letter frequencies
freq <- read.table(
  system.file('scripts', 'letters.txt', package = 'fun'), header = TRUE
)
freq <- freq[order(freq$Letter), ]

set.seed(1)
e2 <- enigma(x, c(3, 2, 2, 3), freq$Count)
e2
enigma(e2)


identical(enigma(e1), enigma(e2))
table(factor(e1, LETTERS))
table(factor(e2, LETTERS))


## the code is every third letter
e3 <- enigma(x, 3)
matrix(e3, 3)[3, ]

}
