% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci's sequence}
\usage{
fibonacci(x)
}
\arguments{
\item{x}{a positive integer}
}
\description{
Calculates the nth number of Fibonacci's sequence.
}
\examples{
fibonacci(20)
sapply(1:20, fibonacci)
}

