% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlspecialchars.R
\name{htmlspecialchars}
\alias{htmlspecialchars}
\title{Replace HTML special characters with HTML entities}
\usage{
htmlspecialchars(string)
}
\arguments{
\item{string}{the string with (or w/o) HTML special chars}
}
\value{
the string with special chars replaced.
}
\description{
The characters \code{c("&", '"', "'", "<", ">")} will be replaced with
\code{c("&amp;", "&quot;", "&#039;", "&lt;", "&gt;")}, respectively.
}
\examples{
htmlspecialchars("<a href = 'https://yihui.org'>Yihui</a>")
# &lt;a href = &#039;https://yihui.org&#039;&gt;Yihui&lt;/a&gt;
}
\references{
\url{https://www.php.net/manual/en/function.htmlspecialchars.php}
}
\seealso{
\code{\link{gsub}}
}
\author{
Yihui Xie <\url{https://yihui.org}>
}
