% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.happy}
\alias{is.happy}
\title{Happy numbers}
\usage{
is.happy(x)
}
\arguments{
\item{x}{a positive integer (any non-digit characters will be ignored)}
}
\description{
For any positive integer, \code{N}, if the sum of the squares of its
digits equals one (or by recursion), then \code{N} is happy \code{:\}}!
Otherwise, \code{N} is sad \code{:c}.
}
\examples{
is.happy(20)
is.happy(19)

}
