% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.prime}
\alias{is.prime}
\title{Prime numbers}
\usage{
is.prime(x)
}
\arguments{
\item{x}{an integer}
}
\description{
Check if a number is prime using a regular expression.
}
\examples{
x <- 1:100
x[is.prime(x)]

}
\references{
\url{http://montreal.pm.org/tech/neil_kandalgaonkar.shtml}
}
\seealso{
\code{\link{decToUna}}
}

