% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nary}
\alias{nary}
\alias{is.unary}
\alias{is.binary}
\alias{binToDec}
\alias{decToBin}
\alias{decToUna}
\title{Nary numbers}
\usage{
is.unary(x)

is.binary(x)

binToDec(x)

decToBin(x)

decToUna(x)
}
\arguments{
\item{x}{a vector to be converted or tested}
}
\description{
Convert to, from, or check if a number is unary or binary.
}
\examples{
x <- 1:100
x_bin <- decToBin(x)
stopifnot(is.binary(x_bin))

x_dec <- binToDec(x_bin)

x_una <- decToUna(x_dec)
stopifnot(is.unary(x_una))

stopifnot(identical(x, x_dec))

}
\seealso{
\code{\link[=intToBits]{rawConversion}}
}
