% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trace_path}
\alias{trace_path}
\title{Trace path}
\usage{
trace_path(lens, turn)
}
\arguments{
\item{lens}{length, number of points}

\item{turn}{turns at each value along \code{lens}}
}
\description{
Trace path
}
\examples{
trace_path(lens = seq(0, 1,  length.out = 200),
           turn = rep(pi/2 * (-1 + 1/200), 200))
trace_path(lens = seq(1, 10, length.out = 1000),
           turn = rep(2 * pi / 10, 1000))
trace_path(lens = seq(0, 1,  length.out = 500),
           turn = seq(0, pi, length.out = 500))
trace_path(lens = seq(0, 1,  length.out = 600) * c(1, -1),
           turn = seq(0, 8*pi, length.out = 600) * seq(-1, 1, length.out = 200))
trace_path(lens = seq(-1, 1,  length.out = 200),
           turn = rep(pi / 1 * (-1 + 1/200), 200))
trace_path(lens = seq(-1, 1,  length.out = 500),
           turn = rep(pi / 1 * (-1 + 1/200), 500))
trace_path(lens = seq(-1, 1,  length.out = 1000),
           turn = rep(pi / 1 * (-1 + 1/500), 1000))

}
