% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchcraft.R
\name{witchcraft}
\alias{witchcraft}
\title{Simple memory game}
\usage{
witchcraft(level = 1, seed)
}
\arguments{
\item{level}{difficulty; choose levels 0 (easy) to 3 (difficult)}

\item{seed}{seed; see \code{\link{set.seed}}}
}
\value{
Interactive game

A prize.
}
\description{
Pick boxes in the correct sequential order to win a prize.
}
\details{
Set the seed to play the same game.
}
\examples{
\dontrun{
witchcraft()
witchcraft(level = 3)
}
}
\author{
Robert Redd
}
