% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bubble_sort}
\alias{bubble_sort}
\title{Bubble sort}
\usage{
bubble_sort(vec)
}
\arguments{
\item{vec}{a vector of unsorted integers}
}
\description{
The bubble sort algorithm iteratively sorts pairs of numbers by swapping
each pair if unsorted and repeating until all numbers are properly sorted.
}
\examples{
\dontrun{
bubble_sort(round(runif(100, 0, 100)))
}
}
\references{
Modified from
\url{http://www.numbertheory.nl/2013/05/10/bubble-sort-implemented-in-pure-r/}
}

