% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ch}
\alias{ch}
\title{Calvin and Hobbes}
\usage{
ch(who)
}
\arguments{
\item{who}{who, \code{'calvin'} or \code{'hobbes'}}
}
\description{
Draw Calvin and Hobbes.
}
\examples{
ch()

}
\seealso{
\url{http://www.stolaf.edu/people/olaf/cs125/}
}

