% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-package.R
\docType{package}
\name{fun-package}
\alias{fun-package}
\alias{fun}
\title{Use R for Fun}
\description{
This is a collection of R games and other funny stuff, such as the classical
Mine sweeper and sliding puzzles.
}
\details{
New games are always welcome; if you know GIT and want to join the
development, please go to \url{https://github.com/yihui/fun}; or simply
contribute ideas at \url{https://github.com/yihui/fun/issues}.
}
\examples{
## See the examples in each function, or check out the demos
demo(package = "fun")
}
\author{
Yihui Xie, Taiyun Wei, and Yixuan Qiu
}
\keyword{package}
