\name{fun-package}
\alias{fun-package}
\alias{fun}
\docType{package}
\title{
Use R for Fun
}
\description{
This is a collection of R games and other funny stuff, such as the classical Mine sweeper and sliding puzzles.
}
\details{
\tabular{ll}{
Package: \tab fun\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2009-04-07\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Yihui Xie, Taiyun Wei and Yixuan Qiu
Maintainer: Yihui Xie <xieyihui@gmail.com>
}
\keyword{ package }
\examples{
# TODO
}
