% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lights_out.R
\name{lights_out}
\alias{lights_out}
\title{Lights out}
\usage{
lights_out(width = 5, height = 5, steps = 3, cheat = FALSE,
           col.off = 'black', col.on = 'white',
           col.frame = 'lightblue', seed)
}
\arguments{
\item{width}{number of lights in x-axis}

\item{height}{number of lights in y-axis}

\item{steps}{number of \code{seed} lights to initialize the puzzle; in
general, the larger \code{steps} is, the more complex the puzzle will be}

\item{cheat}{logical; if \code{TRUE} a data frame indicating the steps to
solve this puzzle will be printed}

\item{col.off}{color when lights off}

\item{col.on}{color when lights on}

\item{col.frame}{color of lights border}

\item{seed}{seed for random number generator}
}
\description{
By default, the white squares in the plot denote lights on, and black
squares are lights off. When you click on a light, this light as well as
its four neighbors will switch their status.

Your mission is to turn off all the lights.
}
\note{
Linux/Mac users have to use \code{X11(type = 'Xlib')} or the
Cairo graphics device \code{Cairo()} in the package
\pkg{cairoDevice}.
}
\examples{
lights_out()
}
\author{
Yixuan Qiu \email{yixuan.qiu@cos.name}
}
\references{
\url{http://en.wikipedia.org/wiki/Lights_Out_(game)}
}

