% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_sweeper.R
\name{mine_sweeper}
\alias{mine_sweeper}
\title{Mine sweeper}
\usage{
mine_sweeper(level = 'beginner', width, height, mines, cheat = FALSE, seed)
}
\arguments{
\item{level}{skill level, one of 'beginner', 'intermediate', 'advanced',
'expert', or 'insane'; custom grids can also be made by supplying all of the
\code{width}, \code{height}, and \code{mines} arguments}

\item{width}{number of grids in horizontal axis}

\item{height}{number of grids in vertical axis}

\item{mines}{number of mines}

\item{cheat}{logical; if \code{TRUE} a matrix indicating the mines will be 
printed}

\item{seed}{seed for random number generator}
}
\description{
The controls should be familiar to you: click the left mouse button to dig 
in an area, and right button to mark or unmark the area with flags.
}
\note{
Linux/Mac users have to use \code{X11(type = 'Xlib')} or the
Cairo graphics device \code{Cairo()} in the package
\pkg{cairoDevice}.
}
\examples{
## a custom grid
mine_sweeper(width = 10, height = 10, mines = 20)

## an impossible grid
mine_sweeper('insane')

## a beginner grid
mine_sweeper()

}
\author{
Yixuan Qiu \email{yixuan.qiu@cos.name}, Robert Redd (modifications)
}
\references{
\url{http://en.wikipedia.org/wiki/Minesweeper_(computer_game)}
}

