% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spark}
\alias{spark}
\alias{sparkbar,}
\title{SparkBar generator}
\usage{
spark(...)
}
\arguments{
\item{...}{numerics}
}
\description{
Generates a sparkbar from a sequence of numbers.
}
\examples{
spark(30, 31, 32, 33)
spark(runif(20))
}
\seealso{
\url{https://gist.github.com/ramnathv/7793167}
}

